add services file

+++ src/main/java/com/model/event/Event.java
@@ -9,7 +9,7 @@ public class Event {
     private LocalDate endDate;
     private String beginLocation;
     private String endLocation;
-    private Double distance;
+    private Float distance;
     private String comment;
     private String status;
     private Boolean hasAccident;
@@ -64,11 +64,11 @@ public class Event {
         this.endLocation = endLocation;
     }
 
-    public Double getDistance() {
+    public Float getDistance() {
         return distance;
     }
 
-    public void setDistance(Double distance) {
+    public void setDistance(Float distance) {
         this.distance = distance;
     }
